services/observe/requirements.txt
PyYAML>=6.0

fastapi>=0.110
uvicorn[standard]>=0.27

# Vision model client (OpenAI-compatible API)
openai>=1.0.0
